+++ src/CBSubmitBannerPopup.hpp
@@ -14,6 +14,8 @@ private:
     void onPickFile(CCObject*);
     void onSubmit(CCObject*);
     void onToggleLimited(CCObject*);
+    void onPreview(CCObject*);
+    void onClosePreview(CCObject*);
 
     geode::TextInput* m_nameInput = nullptr;
     geode::TextInput* m_descInput = nullptr;
@@ -21,6 +23,7 @@ private:
     CCMenuItemToggler* m_limitedToggler = nullptr;
     geode::TextInput* m_amountInput = nullptr;
     CCLabelBMFont* m_fileNameLabel = nullptr;
+    CCMenuItemSpriteExtra* m_previewBtn = nullptr;
 
     std::string m_selectedFilePath;
 };